% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pearls-package.R
\docType{package}
\name{pearls-package}
\alias{pearls}
\alias{pearls-package}
\title{pearls: Operations on Lists of Data Frames}
\description{
One paragraph description of what the package does as one or more full sentences.
}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com}

}
